feat: enhance authentication middleware, improve admin dashboard with statistics, and refactor course and lesson components

+++ src/app/api/category/[id]/route.ts
@@ -4,11 +4,19 @@ import {
   getCategory,
   updateCategory,
 } from "./category.controller";
+import { Roles } from "@/lib/constants";
+import { authMiddleware } from "../../../../../middleware/auth.middleware";
 
 export async function GET(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin, Roles.Student]);
+
+  if (checkAuth.status !== 200) {
+    return checkAuth;
+  }
+
   const { id } = await params;
 
   return getCategory(req, id);
@@ -18,6 +26,12 @@ export async function DELETE(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin]);
+
+  if (checkAuth.status !== 200) {
+    return checkAuth;
+  }
+
   const { id } = await params;
   return deleteCategroy(req, id);
 }
@@ -26,6 +40,11 @@ export async function PUT(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin]);
+
+  if (checkAuth.status !== 200) {
+    return checkAuth;
+  }
   const { id } = await params;
   return updateCategory(req, id);
 }
