+++ src/app/api/enrollment/[id]/route.ts
@@ -4,11 +4,18 @@ import {
   getEnrollment,
   changeEnrollmentStatus,
 } from "../enrollment.controller";
+import { authMiddleware } from "../../../../../middleware/auth.middleware";
+import { Roles } from "@/lib/constants";
 
 export async function GET(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin]);
+
+  if (checkAuth.status !== 200) {
+    return checkAuth;
+  }
   const { id } = await params;
 
   return getEnrollment(req, id);
@@ -18,6 +25,11 @@ export async function DELETE(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin]);
+
+  if (checkAuth.status !== 200) {
+    return checkAuth;
+  }
   const { id } = await params;
   return deleteEnrollment(req, id);
 }
@@ -26,6 +38,12 @@ export async function PATCH(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin, Roles.Student]);
+
+  if (checkAuth.status !== 200) {
+    return checkAuth;
+  }
+
   const { id } = await params;
   return changeEnrollmentStatus(req, id);
 }
