+++ src/app/api/lesson/[id]/route.ts
@@ -1,10 +1,19 @@
 import { NextRequest } from "next/server";
 import { deleteLesson, getLesson, updateLesson } from "../lesson.controller";
+import { authMiddleware } from "../../../../../middleware/auth.middleware";
+import { Roles } from "@/lib/constants";
 
 export async function GET(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin, Roles.Student]);
+
+  if (checkAuth.status !== 200) {
+    
+    return checkAuth;
+  }
+
   const { id } = await params;
 
   return getLesson(req, id);
@@ -14,6 +23,13 @@ export async function DELETE(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin]);
+
+  if (checkAuth.status !== 200) {
+    
+    return checkAuth;
+  }
+
   const { id } = await params;
   return deleteLesson(req, id);
 }
@@ -22,6 +38,13 @@ export async function PATCH(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const checkAuth = await authMiddleware(req, [Roles.Admin]);
+
+  if (checkAuth.status !== 200) {
+    
+    return checkAuth;
+  }
+
   const { id } = await params;
   return updateLesson(req, id);
 }
